requirements.txt
asgiref==3.8.1
Django==5.1.5
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.32.3
gunicorn==22.0.0
whitenoise==6.9.0
dj-database-url==2.3.0